% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_IC.R
\name{calc_IC}
\alias{calc_IC}
\title{Title}
\usage{
calc_IC(Test_Data, factor, return_window = NULL, standarize = FALSE,
  neutrals = NULL, stock_pool = "ZZALL")
}
\arguments{
\item{return_window}{A list. Refer to width in rollapply.}

\item{stock_pool}{}
}
\description{
Title
}
