% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_AR.R
\name{calculate_AR}
\alias{calculate_AR}
\title{Calculate abnormal return by factor model}
\usage{
calculate_AR(data, estimation_window_start, estimation_window_end,
  least_valid_days, estimation_window_start_type = "day",
  estimation_window_end_type = "day", factor_number = 3L)
}
\arguments{
\item{data}{A data frame.

The first column is the trading date whose class is "Date".
The second column is the return.
The others are factors.}

\item{estimation_window_start}{A positive integer giving the start time point of beta estimation window.}

\item{estimation_window_end}{A positive integer giving the end time point of beta estimation window.}

\item{least_valid_days}{A positive integer giving the least vaild trading days in the estimation window.}

\item{estimation_window_start_type}{A character.

"day" means the start time point is the nunmber of days.
"month" means the start time point is the number of months.
"year" means the start time point is the number of years.}

\item{estimation_window_end_type}{A character.

"day" means the end time point is the number of days.
"month" means the end time point is the number of months.
"year" means the end time point is the number of years.}

\item{factor_number}{A positive integer.

1 means there is 1 factor.
3 means there are 3 factors.
4 means there are 4 factors.
5 means there are 5 factors.}
}
\value{
A double vector.
}
\description{
\code{calculate_AR()} calculates abnormal return by factor model.
}
