% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_price_impact_cost.R
\name{calculate_price_impact_cost}
\alias{calculate_price_impact_cost}
\title{Calculate daily price impact cost using wind tick data}
\usage{
calculate_price_impact_cost(tick_data)
}
\arguments{
\item{tick_data}{A data frame.

The tick data come from wind.
This function needs the columns "code", "time", "price", "ask10", "ask9", "ask8", "ask7",
"ask6", "ask5", "ask4", "ask3", "ask2", "ask1", "bid1", "bid2", "bid3", "bid4", "bid5",
"bid6", "bid7", "bid8", "bid9", "bid10", "asize10", "asize9", "asize8", "asize7", "asize6",
"asize5", "asize4", "asize3", "asize2", "asize1", "bsize1", "bsize2", "bsize3", "bsize4",
"bsize5", "bsize6", "bsize7", "bsize8", "bsize9", "bsize10".}
}
\value{
A double matrix.
}
\description{
\code{calculate_price_impact_cost()} calculates daily price impact cost for a stock by using wind tick data
}
