% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnquant.R
\docType{package}
\name{cnquant}
\alias{cnquant}
\alias{cnquant-package}
\title{cnquant: China stock market quant toolbox}
\description{
cnquant aims for quant in China stock market. It provides a comprehensive
toolkbox for quantitative financial analysis as well as high frequency data.
}
\author{
\itemize{
  \item Einzbern LIU \email{ylwpaopao@163.com}
  \item Elf LU
}
}
