% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{df_to_ts}
\alias{df_to_ts}
\title{Transform from data.frame to time series}
\usage{
df_to_ts(df, .datetime = "DATETIME", output_class = "ts",
  frequency = 12L)
}
\arguments{
\item{.datetime}{index.}

\item{output_class}{zoo or ts.}

\item{frequency}{4L or 12L (or NULL) for ts (zoo).}
}
\description{
The data.frame should have one column as index and others as time series.
}
