% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide.R
\name{divide}
\alias{divide}
\title{Divide numeric into quantile groups}
\usage{
divide(x, n_group = 10)
}
\arguments{
\item{x}{A numeric vector.}

\item{n_group}{A positive integer (greater than or equal to 2) giving the
number of quantiles into which x is to be divide.}
}
\value{
An integer vector with the same length of x which takes value from 1:n_group.
}
\description{
\code{divide()} divides numeric into quantiles. It supplements
\code{cut()} by randomize repetitive values into adjacent quantiles.
}
\examples{
divide(rnorm(100))
divide(c(1, 2, 2, 3), n_group = 4)
}
\seealso{
\code{\link[base]{cut}}.
}
