% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind.R
\name{gen_index_dict}
\alias{gen_index_dict}
\title{Generate an index dictionary}
\usage{
gen_index_dict(path, sheet = NULL)
}
\arguments{
\item{sheet}{}
}
\description{
Read in an excel file, and generate an index dictionary as a named chr.
The excel spreadsheet should have 2 columns named code & name respectively.
}
