% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_return.R
\name{layer_return}
\alias{layer_return}
\title{Calculate layer returns from raw stock daily data data frame}
\usage{
layer_return(Test_Data, factor, neutrals = NULL, stock_pool = "ZZALL",
  n_layer = 10L, frequency = 20L, start_point = 1L)
}
\arguments{
\item{Test_Data}{Stock daily data data frame of certain form.}

\item{factor}{The factor variable.}

\item{neutrals}{A character vector specifying style factors (including industry) 
to neutralize upon. Possible components are "CITICS_IND_CODE", "LOG_S_DQ_MV", 
and "LOG_B2M". For example, \code{neutrals = c("CITICS_IND_CODE", "LOG_S_DQ_MV")}.}

\item{stock_pool}{Stock pool in which stocks are selected. "ZZALL", "HS300", or 
"ZZ500". Default to "ZZALL".}

\item{n_layer}{Number of layers. Default 10.}

\item{frequency}{An integer. Adjust position in every how many days?}

\item{start_point}{An integer. Which day from the sample beginning to firstly 
set up position. Usually from 1 to \code{frequency}. Default 1.}
}
\value{
An xts object of returns of each layer.
}
\description{
Calculate layer returns from raw stock daily data data frame
}
