% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_industry.R
\name{match_industry}
\alias{match_industry}
\title{Title}
\usage{
match_industry(.data1, .data2, stock_code, trade_date, ENTRY_DT, REMOVE_DT,
  industry)
}
\arguments{
\item{industry}{}
}
\description{
Title
}
