% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month_end.R
\name{month_end}
\alias{month_end}
\title{Get the last day of month}
\usage{
month_end(x)
}
\arguments{
\item{x}{A 8-digit date character.}
}
\value{
A 8-digit date character.
}
\description{
\code{month_end()} returns the last day of the month.
}
\examples{
month_end("20180712")
}
\seealso{
\code{\link[lubridate]{ceiling_date}}.
}
