% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_para.R
\name{move_para}
\alias{move_para}
\title{Move parallel a matrix and pad NAs}
\usage{
move_para(data, direction = "right", n = 1L)
}
\arguments{
\item{data}{A matrix.}

\item{n}{}
}
\description{
Simple version. Currently only "right" direction is supported.
}
