% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutralize.R
\name{neutralize}
\alias{neutralize}
\title{Neutralize one variable upon others}
\usage{
neutralize(factor, ...)
}
\arguments{
\item{factor}{The variable to neutralize.}

\item{...}{Other variable to neutralize upon.}
}
\value{
The variable after neutralization. Same length of \code{factor}.
}
\description{
Neutralize one variable upon others
}
