% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_tick_time.R
\name{parse_tick_time}
\alias{parse_tick_time}
\title{Parse tick time}
\usage{
parse_tick_time(x)
}
\arguments{
\item{x}{The time character vector variable from Wind tick data format.}
}
\value{
An \code{hms} object.
}
\description{
\code{parse_tick_time()} parses time variable from Wind tick data format.
}
\examples{
parse_tick_time(c("93006000", "103006000"))
}
\seealso{
\code{\link[readr]{parse_time}}.
}
