% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portf_stats.R
\name{portf_stats}
\alias{portf_stats}
\title{Summary statistics for portfolio(s)}
\usage{
portf_stats(Ra, Rb = Ra[, ncol(Ra)], scale = NA, lang = "CHN")
}
\arguments{
\item{Ra}{An xts object of portfolio return.}

\item{Rb}{An xts object of benchmark return.}

\item{scale}{Number of period per year.}

\item{lang}{}
}
\value{
A statistic matrix.
}
\description{
Summary statistics for portfolio(s)
}
