% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tick_data.R
\name{read_tick_data}
\alias{read_tick_data}
\title{Title}
\usage{
read_tick_data(file, col_types = readr::cols_only(time = readr::col_integer(),
  price = readr::col_integer(), volume = readr::col_integer(), turover =
  readr::col_integer(), match_items = readr::col_integer(), bs_flag =
  readr::col_integer(), accvolume = readr::col_integer(), accturover =
  readr::col_integer(), ask10 = readr::col_integer(), ask9 =
  readr::col_integer(), ask8 = readr::col_integer(), ask7 =
  readr::col_integer(), ask6 = readr::col_integer(), ask5 =
  readr::col_integer(), ask4 = readr::col_integer(), ask3 =
  readr::col_integer(),      ask2 = readr::col_integer(), ask1 =
  readr::col_integer(), bid1 = readr::col_integer(), bid2 =
  readr::col_integer(), bid3 = readr::col_integer(), bid4 =
  readr::col_integer(), bid5 = readr::col_integer(), bid6 =
  readr::col_integer(), bid7 = readr::col_integer(), bid8 =
  readr::col_integer(), bid9 = readr::col_integer(), bid10 =
  readr::col_integer(), asize10 = readr::col_integer(), asize9 =
  readr::col_integer(), asize8 = readr::col_integer(), asize7 =
  readr::col_integer(), asize6 = readr::col_integer(),      asize5 =
  readr::col_integer(), asize4 = readr::col_integer(), asize3 =
  readr::col_integer(), asize2 = readr::col_integer(), asize1 =
  readr::col_integer(), bsize1 = readr::col_integer(), bsize2 =
  readr::col_integer(), bsize3 = readr::col_integer(), bsize4 =
  readr::col_integer(), bsize5 = readr::col_integer(), bsize6 =
  readr::col_integer(), bsize7 = readr::col_integer(), bsize8 =
  readr::col_integer(), bsize9 = readr::col_integer(), bsize10 =
  readr::col_integer(), ask_av_price = readr::col_integer(),      bid_av_price =
  readr::col_integer(), total_ask_volume = readr::col_integer(),
  total_bid_volume = readr::col_integer()))
}
\arguments{
\item{col_types}{}
}
\description{
Title
}
