% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_certain.R
\name{replace_certain}
\alias{replace_certain}
\title{Replace a certain value to another}
\usage{
replace_certain(x, pattern, replacement)
}
\arguments{
\item{x}{A vector.}

\item{pattern}{A certain value to be replaced.}

\item{replacement}{A certain value to replace to.}
}
\value{
The modified vector.
}
\description{
\code{replace_certain()} replaces a certain value in a vector to another.
\code{NaN} and \code{NA} are also supported.
}
\examples{
replace_certain(letters, "a", "b")
replace_certain(c(NaN, NA, 3), NaN, 1)
replace_certain(c(NaN, NA, 3), NA, 1)
}
