% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_na.R
\name{show_na}
\alias{show_na}
\title{Show the number of NAs of each column of a data frame}
\usage{
show_na(.data)
}
\arguments{
\item{.data}{A data frame.}
}
\description{
Show the number of NAs of each column of a data frame
}
