% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ_stats.R
\name{summ_stats}
\alias{summ_stats}
\title{Calculate summary statistics}
\usage{
summ_stats(.data, round = NULL, lang = "ENG", ...)
}
\arguments{
\item{...}{}
}
\description{
\code{summ_stats()} calculates frequently used summary statistics for academic paper.
}
