% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_estimate.R
\name{weight_estimate}
\alias{weight_estimate}
\title{Estimate component weight}
\usage{
weight_estimate(df)
}
\arguments{
\item{df}{The first column should be the overall value, and the remained are the components.}
}
\description{
Estimate component weight
}
