% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_mean.R
\name{weighted_mean}
\alias{weighted_mean}
\title{Weighted arithmetic mean}
\usage{
weighted_mean(x, w)
}
\arguments{
\item{x}{A numeric vector.}

\item{w}{A numeric vector of weights the same length as \code{x} giving the
weights.}
}
\value{
A numeric scalar.
}
\description{
\code{weighted_mean()} computes a weighted arithmetic mean. It enhances the
\code{weighted.mean} function with support for \code{NA} values in \code{x}. In
fact, \code{weighted_mean()} excludes element-pair with \coed{NA} either from \coed{x} or
\coed{w}.
}
\examples{
x1 <- c(NA, 2, 3, 4, 5)
w1 <- c(5, NA, 3, 2, 1)
weighted_mean(x1, w1)

x2 <- x1[3:5]
w2 <- w1[3:5]
weighted_mean(x2, w2)
}
