% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winsorize.R
\name{winsorize}
\alias{winsorize}
\title{Winsorize data}
\usage{
winsorize(x, method = "quantile", side = "both", fill = "bound")
}
\arguments{
\item{x}{A numeric vector.}

\item{method}{Default to "quantile". One of "quantile", "median", or "std".
Or a numeric vector of length 2 specify lower and upper bounds.}

\item{side}{Side to winsorize. Default to "both". "both", "right", or "left".}

\item{fill}{Default "bound" to assign outliers to lower and (or) upper bounds, or "NA
to assign outliers to be NA.}
}
\value{
A numeric vector.
}
\description{
\code{winsorize()} winsorizes data to eliminate influence from outliers.
}
\examples{
winsorize(c(1:10, 100))
}
