% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ_stats.R
\name{summ_stats}
\alias{summ_stats}
\title{Calculate summary statistics}
\usage{
summ_stats(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Logical. Should missing values (including \code{NaN}) be removed?}
}
\value{
A numeric vector.
}
\description{
\code{summ_stats()} calculates frequently used summary statistics for academic paper.
}
\examples{
summary_stats(rnorm(100))
}
