% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winsorize.R
\name{winsorize}
\alias{winsorize}
\title{Winsorize data}
\usage{
winsorize(x, method = "median")
}
\arguments{
\item{x}{A numeric vector.}

\item{method}{Default to "median". One of "median", "quantile", or "std".}
}
\value{
A numeric vector.
}
\description{
\code{winsorize()} winsorizes data to eliminate influence from outliers.
}
\examples{
winsorize(c(1:10, 100))
}
